refactor: 미들웨어 폴더 내 유효성 검사 로직만 따로 모아둠
